Move endpoints files
