+++ compiler/server/Server/Endpoints/Project.hs
@@ -1,23 +1,23 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE TypeOperators #-}
 
-module Server.Project
+module Server.Endpoints.Project
   ( projectEndpoints,
     ProjectAPI,
   )
 where
 
 import Servant
-import Server.Project.AddUnitTest
-import Server.Project.BindExpression
-import Server.Project.BindType
-import Server.Project.CreateProject
-import Server.Project.Evaluate
-import Server.Project.Graph
-import Server.Project.ListBindings
-import Server.Project.ListTests
-import Server.Project.Optimise
-import Server.Project.Upgrade
+import Server.Endpoints.Project.AddUnitTest
+import Server.Endpoints.Project.BindExpression
+import Server.Endpoints.Project.BindType
+import Server.Endpoints.Project.CreateProject
+import Server.Endpoints.Project.Evaluate
+import Server.Endpoints.Project.Graph
+import Server.Endpoints.Project.ListBindings
+import Server.Endpoints.Project.ListTests
+import Server.Endpoints.Project.Optimise
+import Server.Endpoints.Project.Upgrade
 import Server.Types
 
 -----
