+++ compiler/server/Server/Servant.hs
@@ -10,10 +10,10 @@ where
 
 import Data.Proxy
 import Servant
-import Server.Compile
-import Server.Expression
-import Server.Project
-import Server.Search
+import Server.Endpoints.Compile
+import Server.Endpoints.Expression
+import Server.Endpoints.Project
+import Server.Endpoints.Search
 import Server.Types
 
 type MimsaAPI = ProjectAPI :<|> SearchAPI :<|> CompileAPI :<|> GetExpression
